// src/app/models/studentUser.ts
export class studentUser
{
    uid?: string;
    email?: string;
    password?: string;
    type?: string;
    name?: string;
    fname?: string;
    number?: string;
    gpa?: number;
    degree?: string;
    city?: string;
    skills?: string;
    gender?: string;
    year?: number;
    schoolName?: string;
    matriculationPassingYear?: string;
    matriculationBoard?:string;
    matriculationSubject?: string;
    universityName?: string;
    dept?:string;
    universityPassingOutYear?:string;
    collegeName?: string;
    IntermediatePassingYear?: string;
    IntermediateBoard?: string;
    IntermediateSubject?: string;
    IntermediatePercentage?: string;
    imageUrl?: string;
    verified?: boolean =false;
}